requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.3
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-guardian==1.4.9
django-taggit==0.23.0
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.8
oauthlib==2.1.0
Pillow==5.4.1
psycopg2==2.7.7
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
urllib3==1.24.1
virtualenv==16.1.0
whitenoise==4.1.2
